backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.1
python-multipart==0.0.6
PyJWT==2.8.0
bcrypt==4.1.2
Pillow==10.1.0
numpy==1.24.4
aiofiles==23.2.1
supabase==2.3.0